Move SierraPhysicalLocationType into the source_model lib
